Simplify toggle-panel-hide script.
This toggles the auto-hide state of each panel, instead of setting all to true if any are false and setting all to false otherwise. This avoids the need for processing the current settings twice and thus eliminates a variable. Plus it's more fun this way, and I still don't care if the behavior is desirable for multiple panels because I use only one.